+++ commands/global/global_commands_router.js
@@ -1,0 +1,8 @@
+import { ping_command } from "./ping/ping.js";
+import "dotenv/config";
+
+export const GLOBAL_COMMANDS = [
+  {
+    ping_command,
+  },
+];
