--- commands/ping/ping.js
@@ -1,17 +1,0 @@
-import { Client, GatewayIntentBits } from 'discord.js';
-const client = new Client({ intents: [GatewayIntentBits.Guilds] });
-
-export const ping_command = [
-    {
-        name: 'ping',
-        description: 'Replies with Pong!',
-    },
-];
-
-client.on('interactionCreate', async interaction => {
-    if (!interaction.isChatInputCommand()) return;
-
-    if (interaction.commandName === 'ping') {
-        await interaction.reply('Pong!');
-    }
-});
