+++ commands/private/private_commands_router.js
@@ -1,0 +1,3 @@
+export const PRIVATE_COMMANDS = [{
+//   smthng
+}]
