+++ logger.js
@@ -1,7 +1,4 @@
 export function Logger(message, client) {
     console.log(`Teste ${message}`);
-    const server = client.guilds.cache.get('1169671375292731452');
-    const channel = server.channels.cache.get('1169671376085454859');
-
-    channel.send(message);
+//     Set log channel details here
 }
